backend/requirements.txt
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
pytest
pytest-django